dont update updated_at when articles in content list are changed, add tests
