+++ features/targeted_publishing/creating_targeted_publishing_destinations_with_content_lists.feature
@@ -13,6 +13,9 @@ Feature: Managing targeted publishing destinations with content lists
       }
     """
     Then the response status code should be 201
+    And the JSON node "contentListItemsUpdatedAt" should be null
+    And the JSON node "updatedAt" should exist
+    And we save it into "updatedAt"
 
     Given I am authenticated as "test.user"
     When I add "Content-Type" header equal to "application/json"
@@ -44,4 +47,19 @@ Feature: Managing targeted publishing destinations with content lists
     And I am authenticated as "test.user"
     And I add "Content-Type" header equal to "application/json"
     Then I send a "GET" request to "/api/v1/content/lists/1/items/"
-    And the Json node "total" should be equal to "1"
+    And the JSON node "total" should be equal to "1"
+
+    And I am authenticated as "test.user"
+    When I add "Content-Type" header equal to "application/json"
+    And I send a "POST" request to "/api/v1/content/push" with body:
+     """
+     {"language": "en", "slugline": "abstract-html-test", "body_html": "<p>some html body</p>", "versioncreated": "2016-09-23T14:57:28+0000", "firstcreated": "2016-09-23T09:11:28+0000", "description_text": "some abstract text", "place": [{"country": "Australia", "world_region": "Oceania", "state": "Australian Capital Territory", "qcode": "ACT", "name": "ACT", "group": "Australia"}], "version": "2", "byline": "ADmin", "keywords": [], "guid": "urn:newsml:sd-master.test.superdesk.org:2022-09-19T09:26:52.402693:f0d01867-e91e-487e-9a50-b638b78fc4bc", "priority": 6, "subject": [{"name": "lawyer", "code": "02002001"}], "urgency": 3, "type": "text", "headline": "Abstract html test updated", "service": [{"name": "Australian General News", "code": "a"}], "description_html": "<p><b><u>some abstract text</u></b></p>", "located": "Sydney", "pubstatus": "usable"}
+     """
+    Then the response status code should be 201
+
+    Given I am authenticated as "test.user"
+    When I add "Content-Type" header equal to "application/json"
+    And I send a "GET" request to "/api/v1/content/lists/1"
+    Then the response status code should be 200
+    And the JSON node "contentListItemsUpdatedAt" should not be null
+    And the JSON node "updatedAt" should be equal to "<<updatedAt>>"
